[#3] Deployed product service & spring cloud config (using hardcoded ip from minikube)
