+++ helpers/commands.md
@@ -19,18 +19,30 @@
   - mvn -Pcentral clean package spring-boot:repackage
   - docker build -t spring-config-server .
 
-#### Run in minikube
+#### Run Spring Cloud Config
   - minikube start -p=3-deploy-product-service --memory=10240 --cpus=4 --disk-size=30g --vm-driver=virtualbox
   - minikube profile 3-deploy-product-service
   - eval $(minikube docker-env) 
     https://stackoverflow.com/a/48999680/1497737
-  - docker build -t spring-config-server /Users/samrat.shaw/Development/github/ms-101/microservices/spring-config-server
+  - docker build -t spring-cloud-config /Users/samrat.shaw/Development/github/ms-101/microservices/spring-cloud-config
   - docker images (should have the built image)
   - eval $(minikube docker-env -u)
   - kubectl create namespace 3-deploy-product-service
   - kubectl config set-context $(kubectl config current-context) --namespace=3-deploy-product-service
-  - kubectl apply -f /Users/samrat.shaw/Development/github/ms-101/kubernetes/3-deploy-product-service/spring-config-server/deployment.yaml
-  - kubectl apply -f /Users/samrat.shaw/Development/github/ms-101/kubernetes/3-deploy-product-service/spring-config-server/service.yaml
+  - kubectl apply -f /Users/samrat.shaw/Development/github/ms-101/kubernetes/3-deploy-product-service/spring-cloud-config/deployment.yaml
+  - kubectl apply -f /Users/samrat.shaw/Development/github/ms-101/kubernetes/3-deploy-product-service/spring-cloud-config/service.yaml
+
+#### Run Product Service
+  - assumptions 
+    - Minikube & spring cloud config is already running
+    - The svc-product is pointing to the correct config server (application.properties)
+    - The application has been already built once.
+    - The correct namespace & context are set in kubectl
+  - eval $(minikube docker-env)
+  - docker build -t svc-product /Users/samrat.shaw/Development/github/ms-101/microservices/svc-product
+  - eval $(minikube docker-env -u)
+  - kubectl apply -f /Users/samrat.shaw/Development/github/ms-101/kubernetes/3-deploy-product-service/svc-product/deployment.yaml
+  - kubectl apply -f /Users/samrat.shaw/Development/github/ms-101/kubernetes/3-deploy-product-service/svc-product/service.yaml
 
 ##### Gotchas
   - Make sure that you build the docker file before deployment. Else the cached build will be containerized.
