+++ helpers/msInitializer.sh
@@ -4,22 +4,22 @@
 #--build=maven \
 #--java-version=11 \
 #--packaging=jar \
-#--name=product-service \
+#--name=svc-product \
 #--package-name=com.samrat.ms.core \
 #--groupId=com.samrat.ms.core \
 #--dependencies=web,actuator,lombok \
 #--version=1.0.0 \
-#product-service
+#svc-product
 
-# Config Server
+## Config Server
 spring init \
 --boot-version=2.3.7 \
 --build=maven \
 --java-version=11 \
 --packaging=jar \
---name=spring-config-server \
+--name=spring-cloud-config \
 --package-name=com.samrat.ms.core \
 --groupId=com.samrat.ms.core \
 --dependencies=web,actuator,cloud-config-server \
 --version=1.0.0 \
-spring-config-server
+spring-cloud-config
