+++ kubernetes/3-deploy-product-service/spring-cloud-config/deployment.yaml
@@ -1,0 +1,20 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: ms-spring-cloud-config
+spec:
+  replicas: 2
+  selector:
+    matchLabels:
+      app: ms-spring-cloud-config
+  template:
+    metadata:
+      labels:
+        app: ms-spring-cloud-config
+    spec:
+      containers:
+        - name: ms-spring-cloud-config
+          image: spring-cloud-config:latest
+          imagePullPolicy: Never
+          ports:
+            - containerPort: 8080
