+++ kubernetes/3-deploy-product-service/spring-cloud-config/service.yaml
@@ -1,0 +1,12 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: service-spring-cloud-config
+spec:
+  type: NodePort
+  selector:
+    app: ms-spring-cloud-config
+  ports:
+    - targetPort: 8080
+      port: 8080
+      nodePort: 30080
