--- kubernetes/3-deploy-product-service/spring-config-server/service.yaml
@@ -1,12 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: ms-spring-config-server-service
-spec:
-  type: NodePort
-  selector:
-    app: ms-spring-config-server
-  ports:
-    - targetPort: 8080
-      port: 8080
-      nodePort: 30080
