+++ kubernetes/3-deploy-product-service/svc-product/service.yaml
@@ -1,0 +1,12 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: service-svc-product
+spec:
+  type: NodePort
+  selector:
+    app: ms-svc-product
+  ports:
+    - targetPort: 8080
+      port: 8080
+      nodePort: 30081
