--- microservices/product-service/Dockerfile
@@ -1,7 +1,0 @@
-FROM openjdk:11.0.9
-
-EXPOSE 8080
-
-ADD ./target/*.jar product-service.jar
-
-ENTRYPOINT ["java","-jar","/product-service.jar"]
