--- microservices/product-service/src/main/java/com/samrat/ms/core/ProductServiceApplication.java
@@ -1,13 +1,0 @@
-package com.samrat.ms.core;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ProductServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ProductServiceApplication.class, args);
-	}
-
-}
