--- microservices/product-service/src/main/java/com/samrat/ms/core/controllers/ProductController.java
@@ -1,20 +1,0 @@
-package com.samrat.ms.core.controllers;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.beans.factory.annotation.Value;
-
-@RestController
-@Slf4j
-public class ProductController {
-
-    @Value("${username}")
-    private String username;
-
-    @GetMapping(value = "/")
-    public String index() {
-        log.info("Received request on index");
-        return "Hello from Product Service. Property from Config Server is = " + username;
-    }
-}
