--- microservices/product-service/src/main/resources/bootstrap.properties
@@ -1,2 +1,0 @@
-spring.application.name=product-service
-spring.cloud.config.url=http://127.0.0.1:8888
