--- microservices/product-service/src/test/java/com/samrat/ms/core/ProductServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.samrat.ms.core;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ProductServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
