+++ microservices/spring-cloud-config/src/main/java/com/samrat/ms/core/SpringCloudConfigApplication.java
@@ -1,0 +1,15 @@
+package com.samrat.ms.core;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.config.server.EnableConfigServer;
+
+@EnableConfigServer
+@SpringBootApplication
+public class SpringCloudConfigApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SpringCloudConfigApplication.class, args);
+	}
+
+}
