+++ microservices/spring-cloud-config/src/main/resources/application.properties
@@ -1,0 +1,12 @@
+
+server.port=8080
+
+# For git repo
+#spring.cloud.config.server.git.uri=https://github.com/samratshaw/ms-101.git
+#spring.cloud.config.server.git.skip-ssl-validation=true
+#management.security.enabled=false
+
+# For File System:
+spring.profiles.active=native
+# spring.cloud.config.server.native.searchLocations=/Users/samrat.shaw/Development/github/ms-101/configurations
+spring.cloud.config.server.native.searchLocations=classpath:/configurations/
