+++ microservices/spring-cloud-config/src/main/resources/configurations/svc-product.properties
