+++ microservices/spring-cloud-config/src/test/java/com/samrat/ms/core/SpringCloudConfigApplicationTests.java
@@ -1,0 +1,13 @@
+package com.samrat.ms.core;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SpringCloudConfigApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
