--- microservices/spring-config-server/src/main/java/com/samrat/ms/core/SpringConfigServerApplication.java
@@ -1,15 +1,0 @@
-package com.samrat.ms.core;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.config.server.EnableConfigServer;
-
-@EnableConfigServer
-@SpringBootApplication
-public class SpringConfigServerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SpringConfigServerApplication.class, args);
-	}
-
-}
