--- microservices/spring-config-server/src/test/java/com/samrat/ms/core/SpringConfigServerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.samrat.ms.core;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpringConfigServerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
