+++ microservices/svc-product/Dockerfile
@@ -1,0 +1,7 @@
+FROM openjdk:11.0.9
+
+EXPOSE 8080
+
+ADD ./target/*.jar svc-product.jar
+
+ENTRYPOINT ["java","-jar","/svc-product.jar"]
