+++ microservices/svc-product/src/main/java/com/samrat/ms/core/SvcProductApplication.java
@@ -1,0 +1,13 @@
+package com.samrat.ms.core;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SvcProductApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SvcProductApplication.class, args);
+	}
+
+}
