+++ microservices/svc-product/src/main/resources/bootstrap.properties
@@ -1,0 +1,6 @@
+spring.application.name=svc-product
+
+spring.cloud.config.uri=http://192.168.99.124:30080
+spring.cloud.config.fail-fast=true
+spring.cloud.config.retry.max-attempts=20
+spring.cloud.config.retry.max-interval=1000
